Feat/demo first example

See merge request papers/airgap/beacon-android-sdk!5

+++ core/src/main/java/it/airgap/beaconsdk/internal/transport/p2p/matrix/store/MatrixStoreAction.kt
@@ -11,7 +11,7 @@ internal sealed class MatrixStoreAction {
 
     data class OnSyncSuccess(
         val syncToken: String?,
-        val pollingTimeout: Int,
+        val pollingTimeout: Long,
         val syncRooms: MatrixSyncResponse.Rooms?,
     ) : MatrixStoreAction()
